+++ src/containerapp/azext_containerapp/_validators.py
@@ -50,9 +50,14 @@ def validate_create(registry_identity, registry_pass, registry_user, registry_se
         raise InvalidArgumentValueError("--registry-identity: expected an ACR registry (*.azurecr.io) for --registry-server")
 
 
-def validate_runtime(runtime, enable_java_metrics):
-    if runtime and runtime.lower() == RUNTIME_GENERIC and enable_java_metrics is not None:
-        raise ValidationError("Not support enable Java metrics with --enable-java-metrics for generic runtime with --runtime generic")
+def validate_runtime(runtime, enable_java_metrics, enable_java_agent):
+    def is_java_enhancement_enabled():
+        return enable_java_agent is not None or enable_java_metrics is not None
+
+    if runtime is None:
+        return
+    if runtime.lower() == RUNTIME_GENERIC and is_java_enhancement_enabled():
+        raise ValidationError("Usage error: --runtime java is required when using --enable-java-metrics or --enable-java-agent")
 
 
 def validate_env_name_or_id(cmd, namespace):
@@ -176,6 +181,7 @@ def validate_build_env_vars(cmd, namespace):
                     env=key_val[0]))
         env_pairs[key_val[0]] = key_val[1]
 
+
 def validate_otlp_headers(cmd, namespace):
     headers = namespace.headers
 
@@ -194,3 +200,9 @@ def validate_otlp_headers(cmd, namespace):
                     header=key_val[0]))
         header_pairs[key_val[0]] = key_val[1]
 
+
+def validate_target_port_range(cmd, namespace):
+    target_port = namespace.target_port
+    if target_port is not None:
+        if target_port < 1 or target_port > 65535:
+            raise ValidationError("Port must be in range [1, 65535].")